requirements.txt
aiogram==3.10.0
aiofiles==23.2.1
aiohttp==3.9.5
python-dotenv==1.0.1